refactor: move owo function to its own file
