+++ fred/libraries/common.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import logging
-import regex as re
 from functools import lru_cache, singledispatch
 from typing import TYPE_CHECKING, Optional
 
@@ -130,32 +129,3 @@ def reduce_str(string: str) -> str:
 
 def mod_name_eq(name1: str, name2: str) -> bool:
     return reduce_str(name1) == reduce_str(name2)
-
-
-owo_table = {
-    r"\bth([aeiou])": r"d\1",
-    r"\bTh([aeiou])": r"D\1",
-    r"oo": r"uwu",
-    r"r": r"w",
-    r"R": r"W",
-    r"ove": r"uv",
-    r"!": r":3",
-    r"(?<![aeiou])([Nn])([aeiou])": r"\1y\2",
-    r"you": "u",
-    r"You": "U",
-    r"fuzzy": r"fuzzy-wuzzy",
-}
-
-
-def owoize(string: str) -> str:
-    new_string: list[str] = []
-    for line in string.split("\n"):
-        new_line: list[str] = []
-        for word in line.split():
-            if re.match(r"://|`", word) is None:
-                for match, sub in owo_table.items():
-                    word = re.sub(match, sub, word)
-            new_line.append(word)
-        new_string.append(" ".join(new_line))
-
-    return "\n".join(new_string)
